Improved error messages and rule file localization
The first parameter is no longer the path to a rule file, but the name
of one of the files in ~/.colorize without extension.  For example,
colorize ~/colorize/rules/gcc.rules becomes colorize gcc.

Omitting the command line argument yields a better error message and
failing to open the rule file does not crash with an exception.